+++ frontend/forms/ResendVerificationEmailForm.php
@@ -1,10 +1,9 @@
 <?php
 
-
-namespace frontend\models;
+namespace frontend\forms;
 
 use Yii;
-use common\models\User;
+use common\entities\User;
 use yii\base\Model;
 
 class ResendVerificationEmailForm extends Model
@@ -25,7 +24,7 @@ class ResendVerificationEmailForm extends Model
             ['email', 'required'],
             ['email', 'email'],
             ['email', 'exist',
-                'targetClass' => '\common\models\User',
+                'targetClass' => '\common\entities\User',
                 'filter' => ['status' => User::STATUS_INACTIVE],
                 'message' => 'There is no user with this email address.'
             ],
