+++ frontend/forms/VerifyEmailForm.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\forms;
 
-use common\models\User;
+use common\entities\User;
 use yii\base\InvalidArgumentException;
 use yii\base\Model;
