+++ frontend/tests/functional/VerifyEmailCest.php
@@ -59,10 +59,10 @@ class VerifyEmailCest
         $I->canSee('Congratulations!', 'h1');
         $I->see('Logout (test.test)', 'form button[type=submit]');
 
-        $I->seeRecord('common\models\User', [
+        $I->seeRecord('common\entities\User', [
            'username' => 'test.test',
            'email' => 'test@mail.com',
-           'status' => \common\models\User::STATUS_ACTIVE
+           'status' => \common\entities\User::STATUS_ACTIVE
         ]);
     }
 }
