+++ frontend/tests/unit/forms/ContactFormTest.php
@@ -1,7 +1,7 @@
 <?php
-namespace frontend\tests\unit\models;
+namespace frontend\tests\unit\forms;
 
-use frontend\models\ContactForm;
+use frontend\forms\ContactForm;
 use yii\mail\MessageInterface;
 
 class ContactFormTest extends \Codeception\Test\Unit
