+++ frontend/tests/unit/forms/PasswordResetRequestFormTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace frontend\tests\unit\models;
+namespace frontend\tests\unit\forms;
 
 use Yii;
-use frontend\models\PasswordResetRequestForm;
+use frontend\forms\PasswordResetRequestForm;
 use common\fixtures\UserFixture as UserFixture;
-use common\models\User;
+use common\entities\User;
 
 class PasswordResetRequestFormTest extends \Codeception\Test\Unit
 {
@@ -19,7 +19,7 @@ class PasswordResetRequestFormTest extends \Codeception\Test\Unit
     {
         $this->tester->haveFixtures([
             'user' => [
-                'class' => UserFixture::className(),
+                'class' => UserFixture::class,
                 'dataFile' => codecept_data_dir() . 'user.php'
             ]
         ]);
